Module google_mybusinessaccountmanagement1::api

Structs

AcceptInvitationRequest
Request message for AccessControl.AcceptInvitation.

Account
An account is a container for your location. If you are the only user who manages locations for your business, you can use your personal Google Account. To share management of locations with multiple users, [create a business account] (https://support.google.com/business/answer/6085339?ref_topic=6085325).

AccountAdminCreateCall
Invites the specified user to become an administrator for the specified account. The invitee must accept the invitation in order to be granted access to the account. See AcceptInvitation to programmatically accept an invitation.

AccountAdminDeleteCall
Removes the specified admin from the specified account.

AccountAdminListCall
Lists the admins for the specified account.

AccountAdminPatchCall
Updates the Admin for the specified Account Admin.

AccountCreateCall
Creates an account with the specified name and type under the given parent. - Personal accounts and Organizations cannot be created. - User Groups cannot be created with a Personal account as primary owner. - Location Groups cannot be created with a primary owner of a Personal account if the Personal account is in an Organization. - Location Groups cannot own Location Groups.

AccountGetCall
Gets the specified account. Returns NOT_FOUND if the account does not exist or if the caller does not have access rights to it.

AccountInvitationAcceptCall
Accepts the specified invitation.

AccountInvitationDeclineCall
Declines the specified invitation.

AccountInvitationListCall
Lists pending invitations for the specified account.

AccountListCall
Lists all of the accounts for the authenticated user. This includes all accounts that the user owns, as well as any accounts for which the user has management rights.

AccountMethods
A builder providing access to all methods supported on account resources. It is not used directly, but through the MyBusinessAccountManagement hub.

AccountPatchCall
Updates the specified business account. Personal accounts cannot be updated using this method.

Admin
An administrator of an Account or a location.

DeclineInvitationRequest
Request message for AccessControl.DeclineInvitation.

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

Invitation
Represents a pending invitation.

ListAccountAdminsResponse
Response message for AccessControl.ListAccountAdmins.

ListAccountsResponse
Response message for Accounts.ListAccounts.

ListInvitationsResponse
Response message for AccessControl.ListInvitations.

ListLocationAdminsResponse
Response message for AccessControl.ListLocationAdmins.

LocationAdminCreateCall
Invites the specified user to become an administrator for the specified location. The invitee must accept the invitation in order to be granted access to the location. See AcceptInvitation to programmatically accept an invitation.

LocationAdminDeleteCall
Removes the specified admin as a manager of the specified location.

LocationAdminListCall
Lists all of the admins for the specified location.

LocationAdminPatchCall
Updates the Admin for the specified location. Only the AdminRole of the Admin can be updated.

LocationMethods
A builder providing access to all methods supported on location resources. It is not used directly, but through the MyBusinessAccountManagement hub.

LocationTransferLocationCall
Moves a location from an account that the user owns to another account that the same user administers. The user must be an owner of the account the location is currently associated with and must also be at least a manager of the destination account.

MyBusinessAccountManagement
Central instance to access all MyBusinessAccountManagement related resource activities

OrganizationInfo
Additional information stored for an organization.

PostalAddress
Represents a postal address, e.g. for postal delivery or payments addresses. Given a postal address, a postal service can deliver items to a premise, P.O. Box or similar. It is not intended to model geographical locations (roads, towns, mountains). In typical usage an address would be created via user input or from importing existing data, depending on the type of process. Advice on address input / editing: - Use an i18n-ready address widget such as https://github.com/google/libaddressinput) - Users should not be presented with UI elements for input or editing of fields outside countries where that field is used. For more guidance on how to use this schema, please see: https://support.google.com/business/answer/6397478

TargetLocation
Represents a target location for a pending invitation.

TransferLocationRequest
Request message for AccessControl.TransferLocation.